backend/requirements.txt
Flask>=3.0.0,<4.0.0
flask-cors>=4.0.0,<5.0.0
python-dotenv>=1.0.0,<2.0.0
bcrypt==5.0.0
PyJWT>=2.8.0,<3.0.0
mysql-connector-python>=9.0.0,<10.0.0
python-dotenv>=1.0.0,<2.0.0
